Fix Boolean Tests
This was working but was changed in February. I'm not confident that the changes
I made here won't have adverse effects elsewhere in the app even though the fit
the specs, which are sensible. If this change isn't made, there is currently no
way to have an invalid boolean field. The inputed value is either tru{e,thy} or
false.

eg.  @boolean.value = "foobar"
     @boolean.value = false